+++ packages/eas-cli/src/utils/__tests__/profiles-test.ts
@@ -1,24 +1,25 @@
 import { Platform } from '@expo/eas-build-job';
+import { errors } from '@expo/eas-json';
 
 import { getProfilesAsync } from '../profiles';
 
 describe(getProfilesAsync, () => {
-  test('defaults to production profile', async () => {
-    const callback = jest.fn();
+  it('defaults to production profile', async () => {
+    const readProfileAsync = jest.fn();
     const result = await getProfilesAsync({
       platforms: [Platform.ANDROID, Platform.IOS],
       profileName: undefined,
-      readProfileAsync: callback,
+      readProfileAsync,
     });
 
     expect(result[0].profileName).toBe('production');
     expect(result[1].profileName).toBe('production');
-    expect(callback).toBeCalledWith(Platform.ANDROID, 'production');
-    expect(callback).toBeCalledWith(Platform.IOS, 'production');
+    expect(readProfileAsync).toBeCalledWith(Platform.ANDROID, 'production');
+    expect(readProfileAsync).toBeCalledWith(Platform.IOS, 'production');
   });
 
-  test('defaults to release profile when production profile is non-existent', async () => {
-    const callback = jest
+  it('defaults to release profile when production profile is non-existent', async () => {
+    const readProfileAsync = jest
       .fn()
       .mockRejectedValueOnce(() => {
         throw new Error();
@@ -30,19 +31,19 @@ describe(getProfilesAsync, () => {
     const result = await getProfilesAsync({
       platforms: [Platform.ANDROID, Platform.IOS],
       profileName: undefined,
-      readProfileAsync: callback,
+      readProfileAsync,
     });
 
     expect(result[0].profileName).toBe('release');
     expect(result[1].profileName).toBe('release');
-    expect(callback).toBeCalledWith(Platform.ANDROID, 'production');
-    expect(callback).toBeCalledWith(Platform.IOS, 'production');
-    expect(callback).toBeCalledWith(Platform.ANDROID, 'release');
-    expect(callback).toBeCalledWith(Platform.IOS, 'release');
+    expect(readProfileAsync).toBeCalledWith(Platform.ANDROID, 'production');
+    expect(readProfileAsync).toBeCalledWith(Platform.IOS, 'production');
+    expect(readProfileAsync).toBeCalledWith(Platform.ANDROID, 'release');
+    expect(readProfileAsync).toBeCalledWith(Platform.IOS, 'release');
   });
 
-  test('fails when neither production or release profiles are present', async () => {
-    const callback = jest
+  it('fails when neither production or release profiles are present', async () => {
+    const readProfileAsync = jest
       .fn()
       .mockRejectedValueOnce(() => {
         throw new Error();
@@ -55,22 +56,36 @@ describe(getProfilesAsync, () => {
       getProfilesAsync({
         platforms: [Platform.ANDROID],
         profileName: undefined,
-        readProfileAsync: callback,
+        readProfileAsync,
       })
     ).rejects.toThrowError(/There is no profile named "production" in eas.json/);
   });
 
-  test('gets a specific profile', async () => {
-    const callback = jest.fn();
+  it('gets a specific profile', async () => {
+    const readProfileAsync = jest.fn();
     const result = await getProfilesAsync({
       platforms: [Platform.ANDROID, Platform.IOS],
       profileName: 'custom-profile',
-      readProfileAsync: callback,
+      readProfileAsync,
     });
 
     expect(result[0].profileName).toBe('custom-profile');
     expect(result[1].profileName).toBe('custom-profile');
-    expect(callback).toBeCalledWith(Platform.ANDROID, 'custom-profile');
-    expect(callback).toBeCalledWith(Platform.IOS, 'custom-profile');
+    expect(readProfileAsync).toBeCalledWith(Platform.ANDROID, 'custom-profile');
+    expect(readProfileAsync).toBeCalledWith(Platform.IOS, 'custom-profile');
+  });
+
+  it('throws validation error if eas.json is invalid', async () => {
+    const readProfileAsync = jest.fn().mockImplementation(() => {
+      throw new errors.InvalidEasJsonError('eas.json is not valid');
+    });
+
+    await expect(
+      getProfilesAsync({
+        platforms: [Platform.ANDROID, Platform.IOS],
+        profileName: undefined,
+        readProfileAsync,
+      })
+    ).rejects.toThrowError(/eas.json is not valid/);
   });
 });